% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1formatRs.R
\name{format_m1}
\alias{format_m1}
\title{Formatter for map1 file}
\usage{
format_m1(path.csv, date1, date2)
}
\arguments{
\item{path.csv}{Path to map1 file location}

\item{date1}{Name of reader start date column}

\item{date2}{Name of reader end date column}
}
\description{
format_m1 formats the map1 file to be compatible with pit_master
}
\details{
This function automatically formats the first map file that contains the 
names, deployment dates, and serial numbers of each reader.
}
