% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1formatRs.R
\name{format_m2}
\alias{format_m2}
\title{formatter for map2 file}
\usage{
format_m2(path.csv, date1, id.colname = id, keep.cols)
}
\arguments{
\item{path.csv}{Path to map2 file location}

\item{date1}{Name of sample date column. Note: the biometrics in this column are what will be associated the pit_ids}

\item{id.colname}{Name of column where PIT tag IDs are stored.}

\item{keep.cols}{list of columns to keep from map2 file i.e., which variables do you want carried over?}
}
\description{
format_m2 formats the map2 file
}
\details{
This function automatically formats the second map file that contains the 
biological information associated with pit_tag IDs.
}
