% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3explorRs.R
\name{inds_by_dates}
\alias{inds_by_dates}
\title{Primary manipulator function to restructure data objects}
\usage{
inds_by_dates(
  x,
  obs.groups = c(id),
  obs.unit = c("weeks", "months", "years"),
  date.colname
)
}
\arguments{
\item{x}{main data object, preferably returned from 'workhorse'}

\item{obs.groups}{list of columns by which observations will be grouped}

\item{obs.unit}{time unit of which to summarize observations}

\item{date.colname}{name of the column containing dates}
}
\description{
Restructure dataframes based on user-defined observation groups and time units
}
\details{
Optional section!
}
\examples{
# ids_wks <- compress_ids_date(x = pm2,
# obs.groups =  c(id, sex, site.x),
# obs.unit = "weeks",
# date.colname = date)
}
