% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2assemblRs.R
\name{integrate_bio}
\alias{integrate_bio}
\title{Primary integration function for biological data}
\usage{
integrate_bio(x, y, joiner = c(), remove.tt = FALSE)
}
\arguments{
\item{x}{data object returned from 'workhorse' that contains tag reads}

\item{y}{data object returned from 'format_m2' that contains biological data}

\item{joiner}{name of shared columns in both x and y dataframes that associates tag reads to biological data (PIT tag ID)}

\item{remove.tt}{drop or keep reads of test tags that are listed in first map file (m1). This defaults to FALSE. If TRUE, test tag reads are ommitted from returned object}
}
\description{
Conveniently joins biological variables
}
\details{
Optional section! Functionally, this is a left_join that joins bio data in m2 file with tag reads, drops test tag reads, and renames columns for clarity.
Currently, this integration does not account for time-varying biological variables (body mass, age, etc)
}
\examples{
# working_pit <- integrate_ids(mastr_pit, m2, 
# joiner = 'pit_id',
# remove.tt = "Y")
}
