% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3explorRs.R
\name{pit_survival}
\alias{pit_survival}
\title{Create survival object for use in 'survival' package}
\usage{
pit_survival(m1, pit_workr, readr_check, min.reads = 2, min.detects = 0)
}
\arguments{
\item{map.file1}{main data object, preferably returned from 'workhorse'}

\item{obs.groups}{list of columns by which observations will be grouped}

\item{obs.unit}{time unit of which to summarize observations}

\item{date.colname}{name of the column containing dates}
}
\description{
Uses several previously built dataframes to create 'sampling data frame' from individuals observations for survival analyses
}
\details{
Optional section!
}
\examples{
# ids_wks <- compress_ids_date(x = pm2,
# obs.groups =  c(id, sex, site.x),
# obs.unit = "weeks",
# date.colname = date)
}
