% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitmastr-package.R
\docType{package}
\name{pitmastr-package}
\alias{pitmastr}
\alias{pitmastr-package}
\title{pitmastr: Pipeline to Wrangle PIT Tag Data}
\description{
Provides a suite of user-friendly functions to assemble, manipulate, and format datasets associated with PIT tag systems.
}
\author{
\strong{Maintainer}: Macy J Kailing \email{macykailing@gmail.com}

Authors:
\itemize{
  \item Joseph R Hoyt
  \item Kate E Langwig
}

}
\keyword{internal}
