% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4system_checkRs.R
\name{readr_check}
\alias{readr_check}
\title{Evaluates system operation}
\usage{
readr_check(x, sys.colname, date1, min.reads = 2)
}
\arguments{
\item{x}{dataframe with all test tag reads}

\item{sys.colname}{name of the column where unique system names are stored (the grouping variable by which to summarize tag reads)}

\item{date1}{date variable to summarize tag reads by}

\item{min.reads}{number of times a test tag should be read in a date to be considered. Defaults to 2.}
}
\description{
Summarizes the number of times a test tag was fired to determine if system was operating each day
}
\details{
this seems optional!
}
\examples{
# TBD
}
