% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1formatRs.R
\name{scale_biodate}
\alias{scale_biodate}
\title{Conversion for dates from calendar days to animal based dates}
\usage{
scale_biodate(x, date.colname = date, set.hour = 7)
}
\arguments{
\item{x}{dataframe where new biodate column will be stored}

\item{date.colname}{name of column containing date of tag reads. Defaults to 'date' as in objects returned from pitmastr functions}

\item{set.hour}{set the hour interval to adjust dates by. Defaults to 7, changing any reads that occur between 12am and 7am to be the previous dates}
}
\description{
Converts dates to begin at a user defined hour, if different than 12am to reflect organisms' biological day-cycle. If tag reads occur between defined hours,
then classify the read as the previous date.
}
\details{
this seems optional!
}
\examples{
# TBD
}
