% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1formatRs.R
\name{format_m1}
\alias{format_m1}
\title{Formatter for map1 file}
\usage{
format_m1(path.csv, date1, date2)
}
\arguments{
\item{path.csv}{Path to map1 file location}

\item{date1}{Name of reader start date column}

\item{date2}{Name of reader end date column}
}
\description{
Formats the first map file that contains the names, deployment dates, and
serial numbers of each reader to be compatible with 'workhorse' function.
}
\details{
this seems optional!
}
\examples{
# m1 <- format_m1(
# path.csv = "/Users/mkailing/Dropbox/Kailing_Projects/pitmastr/inst/dummy_MW/sn_map.csv",
# date1 = start,
# date2 = end)
}
