% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1formatRs.R
\name{format_m2}
\alias{format_m2}
\title{Formatter for map2 file}
\usage{
format_m2(path.csv, date1, id.colname = id, keep.cols)
}
\arguments{
\item{path.csv}{Path to map2 file location}

\item{date1}{Name of sample date column. Note: the biometrics in this column are what will be associated the pit_ids}

\item{id.colname}{Name of column where PIT tag IDs are stored.}

\item{keep.cols}{list of columns to keep from map2 file i.e., which variables do you want carried over?}
}
\description{
Return the second map file that contains the biological variables to link to Tag IDs
in an object formatted to be compatible with an object from 'workhorse'.
}
\details{
this seems optional!
}
\examples{
# m2 <- format_m2(
# path.csv = "/Users/mkailing/Dropbox/Kailing_Projects/pitmastr/inst/dummy_MW/sn_map.csv",
# date1 = date,
# id.colname = pit_id,
# keep.cols = c(sex, species, age, mass, site))
}
