% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4system_checkRs.R
\name{readr_viz}
\alias{readr_viz}
\title{Visualizes system operation}
\usage{
readr_viz(df)
}
\arguments{
\item{df}{dataframe with all test tag reads}
}
\description{
Plots the number of times a test tag was fired, indicating whether reader was operating.
}
\details{
this seems optional!
}
\examples{
# TBD
}
