% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2assemblRs.R
\name{workhorse}
\alias{workhorse}
\title{primary assembler for joining detections}
\usage{
workhorse(directory, string, remove.dup = FALSE, map.file1)
}
\arguments{
\item{directory}{path to directory containing all the files saved from readers}

\item{string}{pattern of characters shared among parent folders containing the files from readers}

\item{remove.dup}{whether duplicate detections (no including reads of test tag) should be dropped; remove.dup defaults to FALSE such that all tag reads are retained}

\item{map.file1}{name of m1, best used with an object returned from the 'format_m1' formatter function.}
}
\description{
workhorse searches across subfolders of a specified directory to extract all tag reads
}
\details{
This function iteratively searches among subfolders for each of the different file types in which tag reads can be stored. Each detection is stored as a row with the respected systems information (serial.number, site, folder location)
stored with the output. ELABORATE ON MORE SPECIFICS!
}
